add missing cas

+++ jdk17u0117/src/hotspot/share/jtsan/jtsanGlobals.cpp
@@ -5,11 +5,11 @@ volatile bool    _is_jtsan_initialized = false;
 volatile uint8_t _gc_epoch = 0;
 
 uint8_t get_gc_epoch(void) {
-    return Atomic::load(&_gc_epoch);
+    return Atomic::load(&(uint8_t)_gc_epoch);
 }
 
 void increment_gc_epoch(void) {
-    Atomic::inc(&_gc_epoch);
+    Atomic::inc(&(uint8_t)_gc_epoch);
 }
 
 bool is_jtsan_initialized(void) {
